break out fingerprinter module
