+++ lib/file-source/src/lib.rs
@@ -5,11 +5,13 @@ extern crate tracing;
 
 mod file_server;
 mod file_watcher;
+mod fingerprinter;
 mod internal_events;
 mod metadata_ext;
 pub mod paths_provider;
 
-pub use self::file_server::{FileServer, Fingerprinter, Shutdown as FileServerShutdown};
+pub use self::file_server::{FileServer, Shutdown as FileServerShutdown};
+pub use self::fingerprinter::Fingerprinter;
 pub use self::internal_events::FileSourceInternalEvents;
 
 type FileFingerprint = u64;
